requirements.txt
ocrmypdf
PyMuPDF
Pillow
